+++ mysql_db_admin.py
@@ -590,8 +590,8 @@ def run_program(args_array, func_dict, **kwargs):
         mongo = gen_libs.load_module(args_array["-m"], args_array["-d"])
 
     if args_array.get("-e", None):
-        mail = setup_mail(args_array.get("-e"),
-                          subj=args_array.get("-s", None))
+        mail = gen_class.setup_mail(args_array.get("-e"),
+                                    subj=args_array.get("-s", None))
 
     # Intersect args_array and func_dict to determine which functions to call.
     for x in set(args_array.keys()) & set(func_dict.keys()):